refactor: Create function to build and setup the argilla server
