+++ tests/unit/server/conftest.py
@@ -13,25 +13,25 @@
 #  limitations under the License.
 
 import contextlib
-from typing import AsyncGenerator, Dict, Generator
+from typing import Dict, Generator
 
 import pytest
 import pytest_asyncio
+from httpx import AsyncClient
+from opensearchpy import OpenSearch
+
 from argilla._constants import API_KEY_HEADER_NAME, DEFAULT_API_KEY
+from argilla.server.app import app
 from argilla.server.daos.backend import GenericElasticEngineBackend
 from argilla.server.daos.datasets import DatasetsDAO
 from argilla.server.daos.records import DatasetRecordsDAO
 from argilla.server.database import get_async_db
 from argilla.server.models import User, UserRole, Workspace
-from argilla.server.search_engine import OpenSearchEngine, SearchEngine, get_search_engine
-from argilla.server.server import app
+from argilla.server.search_engine import SearchEngine, get_search_engine
 from argilla.server.services.datasets import DatasetsService
 from argilla.server.settings import settings
 from argilla.utils import telemetry
 from argilla.utils.telemetry import TelemetryClient
-from httpx import AsyncClient
-from opensearchpy import OpenSearch
-
 from tests.database import TestSession
 from tests.factories import AnnotatorFactory, OwnerFactory, UserFactory
